==> 1.Data_load.ipynb <==
import pymongo
import pandas as pd
import math
from datetime import datetime

# import dask.dataframe as dd ## 대규모 데이터 처리
# import vaex ## 대규모 데이터 처리

import os
from dotenv import load_dotenv

# .env 파일에서 환경 변수 로드
load_dotenv('C:/py_src/awake/env')

# 환경 변수에서 MongoDB 연결 정보 가져오기
mongo_password = os.getenv('MONGO_PASSWORD')
mongo_user = os.getenv('MONGO_USER')

# MongoDB 연결 URL
url = f"mongodb+srv://{mongo_user}:{mongo_password}@meercat-external.udyfs.mongodb.net/?retryWrites=true&w=majority&appName=meercat-external"
client = pymongo.MongoClient(url, serverSelectionTimeoutMS=100000)

# 연결 확인
from pymongo.errors import ServerSelectionTimeoutError

try:
    # 데이터베이스 목록을 가져오는 쿼리
    databases = client.list_database_names()
    print("Connected successfully. Databases:", databases)
    
except ServerSelectionTimeoutError as err:
    print("Connection failed:", err)

# Test DB 선택
db = client.Test

# Test DB 컬렉션 확인
collections = db.list_collection_names()

# 단위 환산
def convert_bytes(num):
    for x in ['bytes', 'KB', 'MB', 'GB', 'TB']:
        if num < 1024.0:
            return f"{num:.2f} {x}"
        num /= 1024.0

for collection_name in collections:
    # 컬렉션 통계 정보 가져오기
    stats = db.command("collStats", collection_name)

    # 컬렉션의 크기와 문서 수 출력    
    print(f"Collection '{collection_name}' size: {convert_bytes(stats['size'])}")
    print(f"Collection '{collection_name}' document count: {stats['count']}")
    print(f"Total index size: {convert_bytes(stats['totalIndexSize'])}")
    print('--------------------------------------------------------------------')

## 인스타 관련 데이터 제외
## 'youtube_videos'/'youtube_datas' 따로 수집
## 'youtube_report_v2'/'youtube_report' 날짜 정보불분명, 다른 테이블 정보와 중복
collections_need = [
    'youtube_revenue',
    'youtube_channel_demographics',
    'youtube_users',
    'youtube_channel_locations',
    'youtube_daily_channel_basics',
    'youtube_subscriber',
    ]

# 데이터 로드
youtube_dict={}
for collection_name in collections_need:

    youtube_dict[collection_name] = pd.DataFrame(list(db[collection_name].find()))
    
    print(collection_name)
    print(youtube_dict[collection_name].columns)
    print(">> Success")
    print("--------------------------------------")
    print("")

# 키값 형식 맞추기
for key, df in youtube_dict.items():
    # 각 딕셔너리의 value가 DataFrame인지 확인
    if isinstance(df, pd.DataFrame):
        # '_id' 컬럼이 있는지 확인하고 문자열로 변환
        if '_id' in df.columns:
            df['_id'] = df['_id'].astype(str)
        
        # 'youtube_user_id' 컬럼이 있는지 확인하고 문자열로 변환
        if 'youtube_user_id' in df.columns:
            df['youtube_user_id'] = df['youtube_user_id'].astype(str)

        # 변경된 DataFrame을 다시 딕셔너리에 저장
        youtube_dict[key] = df

del df

youtube_dict.keys()

### youtube_revenue

youtube_revenue = youtube_dict['youtube_revenue']

len(youtube_revenue['youtube_user_id'].unique())
## 유튜버 계정 : 265개

# 계정별 일일 수익
num_col = youtube_revenue.select_dtypes(['float64','int64']).columns ## 숫자형 컬럼 추출
youtube_revenue = youtube_revenue.groupby(['youtube_user_id','data_created_at'])[num_col].sum().reset_index()
youtube_revenue = youtube_revenue.drop('__v', axis=1)

youtube_revenue

### youtube_report_v2

youtube_report_v2 = youtube_dict['youtube_report_v2']

len(youtube_report_v2['youtube_user_id'].unique())
## 유튜버 계정 : 276개

# del youtube_report_v2

# 필요 컬럼 추출
youtube_report_v2 = youtube_report_v2[['youtube_user_id','content','phone_number','requested']].drop_duplicates().sort_values('youtube_user_id').reset_index(drop=True)

# 날짜형식 변경
youtube_report_v2['date'] = youtube_report_v2['content'].str.split('\n').str[2].str.split(':').str[1].str.strip().str.split(' ').str[:2].apply(''.join)
youtube_report_v2['date'] = youtube_report_v2['date'].apply(lambda x : datetime.strptime(x, '%m월%d일').replace(year=2024))

# 필요 정보 추출
youtube_report_v2['subscriber_count'] = youtube_report_v2['content'].str.split('\n').str[5]
youtube_report_v2['views'] = youtube_report_v2['content'].str.split('\n').str[7]
youtube_report_v2['watched_sum'] = youtube_report_v2['content'].str.split('\n').str[9]
youtube_report_v2['watched_avg'] = youtube_report_v2['content'].str.split('\n').str[11]

youtube_report_v2 = youtube_report_v2.drop('content',axis=1)

# 최종 데이터셋 생성
youtube_report_v2 = youtube_report_v2[['youtube_user_id', 'date', 'phone_number', 'subscriber_count', 'views', 'watched_sum', 'watched_avg', 'requested']]
youtube_report_v2 = youtube_report_v2.sort_values(['youtube_user_id','date']).drop_duplicates().reset_index(drop=True)

### youtube_channel_demographics
- 유튜버계정 일일 연령대, 성별 분포
- demographics 컬럼 분포 합 100%

youtube_channel_demographics = youtube_dict['youtube_channel_demographics']

len(youtube_channel_demographics['youtube_user_id'].unique())
## 유튜버 계정 : 905개

# 필요 컬럼 추출
youtube_channel_demographics = youtube_channel_demographics[~youtube_channel_demographics['demographics'].isnull()].sort_values(['youtube_user_id','end_date']).reset_index(drop=True)
youtube_channel_demographics = youtube_channel_demographics[['youtube_user_id', 'end_date', 'demographics']]

# 최종 데이터셋 생성
youtube_channel_demographics = pd.concat([youtube_channel_demographics, pd.json_normalize(youtube_channel_demographics['demographics'])],axis=1)
youtube_channel_demographics = youtube_channel_demographics.drop('demographics',axis=1)
youtube_channel_demographics = youtube_channel_demographics[youtube_channel_demographics.columns[:16]]
youtube_channel_demographics = youtube_channel_demographics.sort_values(['youtube_user_id', 'end_date']).drop_duplicates().reset_index(drop=True)

### youtube_report

youtube_report = youtube_dict['youtube_report']

len(youtube_report['youtube_user_id'].unique())
## 유튜버 계정 : 194개

youtube_report = youtube_report[['youtube_user_id','contents','phone_number','request']].drop_duplicates().sort_values('youtube_user_id').reset_index(drop=True)

# 날짜형식 변경
youtube_report['date'] = youtube_report['contents'].str.split('\n').str[2].str.split(':').str[1].str.strip().str.split(' ').str[:2].apply(''.join)
youtube_report['date'] = youtube_report['date'].apply(lambda x : datetime.strptime(x, '%m월%d일').replace(year=2024))

# 필요 정보 추출
youtube_report['subscriber_count'] = youtube_report['contents'].str.split('\n').str[5]
youtube_report['views'] = youtube_report['contents'].str.split('\n').str[7]
youtube_report['watched_sum'] = youtube_report['contents'].str.split('\n').str[9]
youtube_report['watched_avg'] = youtube_report['contents'].str.split('\n').str[11]

youtube_report = youtube_report.drop('contents',axis=1)

# 최종 데이터셋 생성
youtube_report = youtube_report[['youtube_user_id', 'date', 'phone_number', 'subscriber_count', 'views', 'watched_sum', 'watched_avg', 'request']]
youtube_report = youtube_report.sort_values(['youtube_user_id','date']).drop_duplicates().reset_index(drop=True)

### youtube_users

youtube_users = youtube_dict['youtube_users']

len(youtube_users['channel_id'].unique())
## 유튜버 계정 : 883개

# 필요컬럼추출
youtube_users = youtube_users[['channel_id', 'channel_title', 'phone_num', 'report_user_id', 'statistics','published_at']] 
## published_at : 유튜브 가입일
## 'statistics' 컬럼의 'subscriberCount' 정보와 'subscriber_count' 컬럼 정보가 다름 --> 구독자 수
## 'channel_id' 컬럼, 'contentDetails' 컬럼의 'uploads' 같은 정보

# cast 하여 최종 데이터셋 생성
youtube_users = pd.concat([youtube_users, pd.json_normalize(youtube_users['statistics'])],axis=1)
youtube_users = youtube_users.drop(['statistics','hiddenSubscriberCount'],axis=1)
youtube_users = youtube_users.dropna(how = 'all')

# null 값 0으로 대체
youtube_users[['viewCount', 'subscriberCount', 'videoCount']] = youtube_users[['viewCount', 'subscriberCount', 'videoCount']].fillna(0)

youtube_users = youtube_users.sort_values('channel_id').drop_duplicates().reset_index(drop=True)

### youtube_channel_locations
- 채널 구독자 위치

youtube_channel_locations = youtube_dict['youtube_channel_locations']

len(youtube_channel_locations['youtube_user_id'].unique())
## 유튜버 계정 : 906개

# 필요컬럼추출
youtube_channel_locations = youtube_channel_locations[youtube_channel_locations['locations'].apply(lambda x: len(x) > 0)]
youtube_channel_locations = youtube_channel_locations[['youtube_user_id','end_date','locations']]

# melt, cast 하여 최종데이터셋 생성
youtube_channel_locations = youtube_channel_locations.explode(['locations']).reset_index(drop=True)

# cast 하여 최종 데이터셋 생성
youtube_channel_locations = pd.concat([youtube_channel_locations,pd.json_normalize(youtube_channel_locations['locations'])], axis=1)
youtube_channel_locations = youtube_channel_locations.drop(['locations','subscribersGained','subscribersLost'],axis=1)
youtube_channel_locations = youtube_channel_locations[youtube_channel_locations[youtube_channel_locations.columns[3:]].apply(sum,axis=1)!=0] ## 모두 0인 행 제거
youtube_channel_locations = youtube_channel_locations.drop_duplicates().sort_values(['youtube_user_id','end_date']).reset_index(drop=True)

### youtube_daily_channel_basics

youtube_daily_channel_basics = youtube_dict['youtube_daily_channel_basics']

len(youtube_daily_channel_basics['youtube_user_id'].unique())
## 유튜버 계정 : 906개

# 필요컬럼추출
youtube_daily_channel_basics = youtube_daily_channel_basics[youtube_daily_channel_basics['daily_basics'].apply(lambda x: len(x) > 0)]
youtube_daily_channel_basics = youtube_daily_channel_basics[['youtube_user_id','daily_basics']] ## 'end_date' --> 'day' 컬럼 활용
youtube_daily_channel_basics = youtube_daily_channel_basics.explode(['daily_basics']).reset_index(drop=True) ## melt

# daily_basics 컬럼 형식 맞추기 --> 딕셔너리 값들 cast 형태로 맞추기
youtube_daily_channel_basics_cast = []
for item in youtube_daily_channel_basics['daily_basics']:
    if isinstance(item, list):
        youtube_daily_channel_basics_cast.extend(item)
    else:
        youtube_daily_channel_basics_cast.append(item)

youtube_daily_channel_basics_cast = pd.json_normalize(youtube_daily_channel_basics_cast)

# melt, cast 하여 최종 데이터셋 생성
youtube_daily_channel_basics = pd.concat([youtube_daily_channel_basics, youtube_daily_channel_basics_cast],axis=1)
youtube_daily_channel_basics = youtube_daily_channel_basics.drop('daily_basics',axis=1)
youtube_daily_channel_basics = youtube_daily_channel_basics.fillna(0) ## null값 0으로 대체
youtube_daily_channel_basics = youtube_daily_channel_basics[youtube_daily_channel_basics[youtube_daily_channel_basics.columns[3:]].sum(axis=1)!=0] ## 모두 0인 행 제거
youtube_daily_channel_basics = youtube_daily_channel_basics.sort_values(['youtube_user_id','day']).reset_index(drop=True)

del youtube_daily_channel_basics_cast

### youtube_subscriber

youtube_subscriber = youtube_dict['youtube_subscriber']

len(youtube_subscriber['youtube_user_id'].unique())
## 유튜버 계정 : 994개

# 필요컬럼추출
youtube_subscriber = youtube_subscriber[['youtube_user_id','data_created_at','subscribers_gained','subscribers_lost','subscribers_count']]
youtube_subscriber = youtube_subscriber.drop_duplicates()

youtube_subscriber = youtube_subscriber[youtube_subscriber[youtube_subscriber.columns[2:]].sum(axis=1)!=0] ## 모두 0인 행 제거
youtube_subscriber = youtube_subscriber.sort_values(['youtube_user_id','data_created_at']).reset_index(drop=True)

# 불필요 데이터 삭제 - 메모리 최소화
del youtube_dict

# 데이터 필터링 기준
# 날짜 기준
first_date = datetime(2023, 3, 26)
last_date = datetime(2024, 5, 3)

# 최종데이터셋 기준 필요 계정 수 추출 - report, report_v2 제외
youtube_user_id_outer = list(set(list(youtube_revenue['youtube_user_id']) +
                                #  list(youtube_report_v2['youtube_user_id']) +
                                 list(youtube_channel_demographics['youtube_user_id']) +
                                #  list(youtube_report['youtube_user_id']) +
                                 # list(youtube_users['youtube_user_id']) +
                                 list(youtube_channel_locations['youtube_user_id']) +
                                 list(youtube_daily_channel_basics['youtube_user_id']) +
                                 list(youtube_subscriber['youtube_user_id']) +
                                 list(youtube_revenue['youtube_user_id'])))
print('youtube_user_id_outer', len(youtube_user_id_outer))
## youtube_user_id 모두 포함 912개

youtube_user_id_inner = list(set(youtube_revenue['youtube_user_id']) &
                            #  set(youtube_report_v2['youtube_user_id']) &
                             set(youtube_channel_demographics['youtube_user_id']) &
                            #  set(youtube_report['youtube_user_id']) &
                            #  set(youtube_users['youtube_user_id']) &
                             set(youtube_channel_locations['youtube_user_id']) &
                             set(youtube_daily_channel_basics['youtube_user_id']) &
                             set(youtube_subscriber['youtube_user_id']) &
                             set(youtube_revenue['youtube_user_id']))

print('youtube_user_id_inner',len(youtube_user_id_inner))
## youtube_user_id  공통 포함 250개

### 대용량 테이블 로드
- youtube_videos
- youtube_datas

#### youtube_videos
- 계정별 일일 콘텐츠 정보

collection = db['youtube_videos']  # 컬렉션 선택

# 파이프라인 정의
pipeline = [
    {
        "$match": {
            "youtube_user_id": {
                "$in": youtube_user_id_inner
            },
            "videos": {"$ne": []},
            "end_date": {
                "$gte": first_date,
                "$lte": last_date
            }
        }
    },
    {
        "$sort": {
            "youtube_user_id": 1,
            "end_date": 1
        }
    },
    {
        "$project": {
            "youtube_user_id": 1,
            "end_date": 1,
            "videos": 1
        }
    }
]

# 파이프라인 실행
result = list(collection.aggregate(pipeline, allowDiskUse=True))

# 결과를 Pandas 데이터프레임으로 변환
youtube_videos = pd.DataFrame(result)

len(youtube_videos['youtube_user_id'].unique())

# melt, cast하여 최종데이터셋 생성
# melt
youtube_videos = youtube_videos.explode(['videos']).reset_index(drop=True)

#cast
youtube_videos = pd.concat([youtube_videos, pd.json_normalize(youtube_videos['videos'])],axis=1)

youtube_videos = youtube_videos.drop(['_id','videos'],axis=1)
youtube_videos = youtube_videos.fillna(0) ## null값 0으로 대체
youtube_videos = youtube_videos[youtube_videos[youtube_videos.columns[4:]].sum(axis=1)!=0] ## 모두 0인 행 제거
youtube_videos = youtube_videos.drop_duplicates().sort_values(['youtube_user_id', 'end_date']).reset_index(drop=True)

# 내보내기
# youtube_videos.to_csv('C:/py_src/awake/data/youtube_videos.csv',encoding='utf-8-sig',index=False)

#### youtube_datas

collection = db['youtube_datas']  # 컬렉션 선택

# 파이프라인 정의
pipeline = [
    {
        "$match": {
            "youtube_user_id": {
                "$in": youtube_user_id_inner
            },
            # "end_date": {
            # "$gte": first_date,
            # "$lte": last_date
            # }
        }
    },
    {
        "$sort": {
            "youtube_user_id": 1,
            "data_created_at": 1
        }
    },
    {
        "$project": {
            'youtube_user_id' : 1, 
            'data_created_at' : 1, 
            'published_at' : 1, 
            'channel_id' : 1, 
            'channel_title' : 1, 
            'traffic_source_type' : 1, 
            'yt_search_keyword' : 1, 
            'subscribed_status' : 1,
            'subscriber_count' : 1,
            'video_count' : 1,
            'view_count' : 1,
            'comment_count' : 1,
            'like_count' : 1,
            'dislike_count' : 1,
            'estimated_minutes_watched' : 1,
            'average_view_duration' : 1,
            'status_code' : 1,
            'red_view_count' : 1
        }
    }
]

# 파이프라인 실행
result = list(collection.aggregate(pipeline, allowDiskUse=True))

# 결과를 Pandas 데이터프레임으로 변환
youtube_datas = pd.DataFrame(result)

len(youtube_datas['youtube_user_id'].unique())

# 컬럼 순서 정리
need_col = ['youtube_user_id', 'data_created_at', 'published_at', 'channel_id', 'channel_title', 'traffic_source_type', 'yt_search_keyword', 'subscribed_status','subscriber_count',
            'video_count','view_count','comment_count','like_count','dislike_count','estimated_minutes_watched','average_view_duration','status_code','red_view_count']
youtube_datas = youtube_datas[need_col]

# cast 하여 데이터셋 생성
youtube_datas = pd.concat([youtube_datas, pd.json_normalize(youtube_datas['traffic_source_type']), pd.json_normalize(youtube_datas['subscribed_status'])],axis=1)
youtube_datas = youtube_datas.drop(['traffic_source_type','subscribed_status'],axis=1)

youtube_datas[youtube_datas.columns[6:]] = youtube_datas[youtube_datas.columns[6:]].fillna(0) # null값 0으로 대체
youtube_datas = youtube_datas[youtube_datas[youtube_datas.columns[6:]].sum(axis=1)!=0] ## 모두 0인 행 제거

youtube_datas = youtube_datas.sort_values(['youtube_user_id','data_created_at']).reset_index(drop=True)

# 날짜 기준 필터링
opt_date = (youtube_datas['data_created_at'] >= first_date) & (youtube_datas['data_created_at'] <= last_date)
youtube_datas = youtube_datas[opt_date].sort_values(['youtube_user_id', 'data_created_at']).reset_index(drop=True)
youtube_datas

# 내보내기
# df.to_csv('')
