==> 2.Data_preprocessing.ipynb <==
import pymongo
import pandas as pd
import math
from datetime import datetime

# import dask.dataframe as dd ## 대규모 데이터 처리
# import vaex ## 대규모 데이터 처리

import os
from dotenv import load_dotenv

### 분석용데이터셋 생성
- youtube_user_id_inner 활용
- 날짜 맞춰서 테이블매핑 : 2023-03-26 ~ 2024-05-03

(1) 계정 분석 데이터셋
(2) 콘텐츠 분석 데이터셋

# 데이터 필터링 기준
# 날짜 기준
first_date = datetime(2023, 3, 26)
last_date = datetime(2024, 5, 3)

file_path = 'C:/py_src/awake/data/'

youtube_user_id_inner = pd.read_csv(file_path + 'youtube_user_id_inner.csv')
youtube_user_id_inner = list(youtube_user_id_inner['youtube_user_id'])

youtube_channel_demographics = pd.read_csv(file_path + 'youtube_channel_demographics.csv')

youtube_channel_locations = pd.read_csv(file_path + 'youtube_channel_locations.csv')
youtube_channel_locations = youtube_channel_locations.groupby(['youtube_user_id', 'end_date']).agg({'views': 'sum',
                                                                                                    'estimatedMinutesWatched': 'sum',
                                                                                                    'averageViewDuration': 'mean',
                                                                                                    'averageViewPercentage': 'mean'}).reset_index()

youtube_daily_channel_basics = pd.read_csv(file_path + 'youtube_daily_channel_basics.csv')
youtube_datas = pd.read_csv(file_path + 'youtube_datas.csv')
youtube_revenue = pd.read_csv(file_path + 'youtube_revenue.csv')
youtube_subscriber = pd.read_csv(file_path + 'youtube_subscriber.csv')
youtube_users = pd.read_csv(file_path + 'youtube_users.csv')
youtube_videos = pd.read_csv(file_path + 'youtube_videos.csv')

#### 테이블 현황 및 정리

# 계정 기준 필터링
youtube_revenue = youtube_revenue[youtube_revenue['youtube_user_id'].isin(youtube_user_id_inner)].sort_values(['youtube_user_id','data_created_at']).reset_index(drop=True)

# 날짜 기준 필터링
youtube_revenue['data_created_at'] = pd.to_datetime(youtube_revenue['data_created_at'])
opt_date = (youtube_revenue['data_created_at'] >= first_date) & (youtube_revenue['data_created_at'] <= last_date)
youtube_revenue = youtube_revenue[opt_date].sort_values(['youtube_user_id', 'data_created_at']).reset_index(drop=True)

print(youtube_revenue.columns)
print(youtube_revenue.shape)
print(youtube_revenue['data_created_at'].min(), youtube_revenue['data_created_at'].max())

# youtube_revenue.to_csv('C:/py_src/awake/data/youtube_revenue.csv',encoding='utf-8-sig',index=False)

# 계정 기준 필터링
youtube_channel_demographics = youtube_channel_demographics[youtube_channel_demographics['youtube_user_id'].isin(youtube_user_id_inner)].sort_values(['youtube_user_id','end_date']).reset_index(drop=True)

# 날짜 기준 필터링
youtube_channel_demographics['end_date'] = pd.to_datetime(youtube_channel_demographics['end_date'])
opt_date = (youtube_channel_demographics['end_date'] >= first_date) & (youtube_channel_demographics['end_date'] <= last_date)
youtube_channel_demographics = youtube_channel_demographics[opt_date].sort_values(['youtube_user_id', 'end_date']).reset_index(drop=True)

print(youtube_channel_demographics.columns)
print(youtube_channel_demographics.shape)
print(youtube_channel_demographics['end_date'].min(), youtube_channel_demographics['end_date'].max())

# youtube_channel_demographics.to_csv('C:/py_src/awake/data/youtube_channel_demographics.csv',encoding='utf-8-sig',index=False)

print(youtube_users.columns)
print(youtube_users.shape)

# youtube_users.to_csv('C:/py_src/awake/data/youtube_users.csv',encoding='utf-8-sig',index=False)

# 계정 기준 필터링
youtube_channel_locations = youtube_channel_locations[youtube_channel_locations['youtube_user_id'].isin(youtube_user_id_inner)].sort_values(['youtube_user_id','end_date']).reset_index(drop=True)

# 날짜 기준 필터링
youtube_channel_locations['end_date'] = pd.to_datetime(youtube_channel_locations['end_date'])
opt_date = (youtube_channel_locations['end_date'] >= first_date) & (youtube_channel_locations['end_date'] <= last_date)
youtube_channel_locations = youtube_channel_locations[opt_date].sort_values(['youtube_user_id', 'end_date']).reset_index(drop=True)

print(youtube_channel_locations.columns)
print(youtube_channel_locations.shape)
print(youtube_channel_locations['end_date'].min(), youtube_channel_locations['end_date'].max())

# youtube_channel_locations.to_csv('C:/py_src/awake/data/youtube_channel_locations.csv',encoding='utf-8-sig',index=False)

# 계정 기준 필터링
youtube_daily_channel_basics = youtube_daily_channel_basics[youtube_daily_channel_basics['youtube_user_id'].isin(youtube_user_id_inner)].sort_values(['youtube_user_id','day']).reset_index(drop=True)

# 날짜 기준 필터링
youtube_daily_channel_basics['day'] = pd.to_datetime(youtube_daily_channel_basics['day'])
opt_date = (youtube_daily_channel_basics['day'] >= first_date) & (youtube_daily_channel_basics['day'] <= last_date)
youtube_daily_channel_basics = youtube_daily_channel_basics[opt_date].sort_values(['youtube_user_id', 'day']).reset_index(drop=True)

print(youtube_daily_channel_basics.columns)
print(youtube_daily_channel_basics.shape)
print(youtube_daily_channel_basics['day'].min(), youtube_daily_channel_basics['day'].max())

# youtube_daily_channel_basics.to_csv('C:/py_src/awake/data/youtube_daily_channel_basics.csv',encoding='utf-8-sig',index=False)

# 계정 기준 필터링
youtube_subscriber = youtube_subscriber[youtube_subscriber['youtube_user_id'].isin(youtube_user_id_inner)].sort_values(['youtube_user_id','data_created_at']).reset_index(drop=True)

# 날짜 기준 필터링
youtube_subscriber['data_created_at'] = pd.to_datetime(youtube_subscriber['data_created_at'])
opt_date = (youtube_subscriber['data_created_at'] >= first_date) & (youtube_subscriber['data_created_at'] <= last_date)
youtube_subscriber = youtube_subscriber[opt_date].sort_values(['youtube_user_id', 'data_created_at']).reset_index(drop=True)

print(youtube_subscriber.columns)
print(youtube_subscriber.shape)
print(youtube_subscriber['data_created_at'].min(), youtube_subscriber['data_created_at'].max())

# youtube_subscriber.to_csv('C:/py_src/awake/data/youtube_subscriber.csv',encoding='utf-8-sig',index=False)

# 계정 기준 필터링
youtube_datas = youtube_datas[youtube_datas['youtube_user_id'].isin(youtube_user_id_inner)].sort_values(['youtube_user_id','data_created_at']).reset_index(drop=True)

# 날짜 기준 필터링
youtube_datas['data_created_at'] = pd.to_datetime(youtube_datas['data_created_at'])
opt_date = (youtube_datas['data_created_at'] >= first_date) & (youtube_datas['data_created_at'] <= last_date)
youtube_datas = youtube_datas[opt_date].sort_values(['youtube_user_id', 'data_created_at']).reset_index(drop=True)

print(youtube_datas.columns)
print(youtube_datas.shape)
print(youtube_datas['data_created_at'].min(), youtube_datas['data_created_at'].max())

# youtube_datas.to_csv('C:/py_src/awake/data/youtube_datas.csv',encoding='utf-8-sig',index=False)

# 날짜 기준 필터링
youtube_videos = youtube_videos[youtube_videos['youtube_user_id'].isin(youtube_user_id_inner)].sort_values(['youtube_user_id','end_date']).reset_index(drop=True)

# 날짜 기준 필터링
youtube_videos['end_date'] = pd.to_datetime(youtube_videos['end_date'])
opt_date = (youtube_videos['end_date'] >= first_date) & (youtube_videos['end_date'] <= last_date)
youtube_videos = youtube_videos[opt_date].sort_values(['youtube_user_id', 'end_date']).reset_index(drop=True)

print(youtube_videos.columns)
print(youtube_videos.shape)
print(youtube_videos['end_date'].min(), youtube_videos['end_date'].max())

# youtube_videos.to_csv('C:/py_src/awake/data/youtube_videos.csv',encoding='utf-8-sig',index=False)

#### 계정 분석 데이터셋

youtube_revenue = youtube_revenue.rename(columns={'data_created_at':'date'})
youtube_revenue['date'] = youtube_revenue['date'].astype(str)
youtube_revenue.head()

youtube_channel_demographics = youtube_channel_demographics.rename(columns={'end_date':'date'})
youtube_channel_demographics['date'] = youtube_channel_demographics['date'].astype(str)
youtube_channel_demographics.head()

# 핸드폰 번호 형식으로 변환하는 함수
def format_phone_number(x):
    if pd.isna(x):
        return "010-0000-0000"  # NaN 값을 기본 전화번호 형식으로 대체
    else:
        # 숫자를 정수로 변환한 후 문자열로 변환
        phone_str = str(int(x))
        phone_format = '0' + phone_str

        phone_format = phone_format[:3] + '-' + phone_format[3:7]+ '-' + phone_format[7:]

        
        # 3-3-4 형식으로 포맷팅 (예: 010-1234-5678)
        return phone_format
    
youtube_users['phone_num'] = youtube_users['phone_num'].apply(format_phone_number)
youtube_users.head()

youtube_channel_locations = youtube_channel_locations.rename(columns={'end_date':'date'})
youtube_channel_locations['date'] = youtube_channel_locations['date'].astype(str)
youtube_channel_locations.head()

youtube_daily_channel_basics = youtube_daily_channel_basics.rename(columns={'day':'date'})
youtube_daily_channel_basics['date'] = youtube_daily_channel_basics['date'].astype(str)
youtube_daily_channel_basics.head()

youtube_subscriber = youtube_subscriber.rename(columns={'data_created_at':'date'})
youtube_subscriber['date'] = youtube_subscriber['date'].astype(str)
youtube_subscriber.head()

youtube_datas = youtube_datas.rename(columns={'data_created_at':'date'})
youtube_datas['date'] = youtube_datas['date'].astype(str)
youtube_datas.head()

# 계정 분석 데이터 테이블
# youtube_revenue
# youtube_channel_demographics
# youtube_users
# youtube_channel_locations
# youtube_daily_channel_basics
# youtube_subscriber
# youtube_datas

# # 계정 분석 데이터 테이블 매핑
merge_df_users = pd.merge(youtube_users,youtube_datas,how='outer',on=['channel_id']).reset_index(drop=True)
merge_df_users = pd.merge(merge_df_users,youtube_revenue,how='outer',on=['youtube_user_id','date']).reset_index(drop=True)
merge_df_users = pd.merge(merge_df_users,youtube_channel_demographics,how='outer',on=['youtube_user_id','date']).reset_index(drop=True)
merge_df_users = pd.merge(merge_df_users,youtube_channel_locations,how='outer',on=['youtube_user_id','date']).reset_index(drop=True)
merge_df_users = pd.merge(merge_df_users,youtube_daily_channel_basics,how='outer',on=['youtube_user_id','date']).reset_index(drop=True)
merge_df_users = pd.merge(merge_df_users,youtube_subscriber,how='outer',on=['youtube_user_id','date']).reset_index(drop=True)
# merge_df_users = pd.merge(merge_df_users,youtube_datas,how='outer',on=['youtube_user_id','date']).reset_index(drop=True)
# merge_df_users = pd.merge(merge_df_users,youtube_users,how='outer',on=['channel_id']).reset_index(drop=True)

merge_df_users = merge_df_users.drop(['channel_title_y','published_at_y'],axis=1)
merge_df_users

# # 테이블별 필요 컬럼
# 'youtube_user_id', 'date', 
# 'channel_id', 'channel_title_x', 'phone_num', 'report_user_id', 'published_at_x', 'viewCount', 'subscriberCount', 'videoCount',   ## youtube_users

# 'yt_search_keyword','UNSUBSCRIBED', 'SUBSCRIBED', ## youtube_datas

# 'estimated_revenue','estimated_ad_revenue', 
# 'estimated_red_partner_revenue',
# 'gross_revenue', 'cpm_x', 'ad_impressions', 
# 'monetized_playbacks','playback_based_cpm', ## youtube_revenue

# 'age13-17.female', 'age13-17.male',
# 'age18-24.female', 'age18-24.male', 'age25-34.female', 'age25-34.male',
# 'age35-44.female', 'age35-44.male', 'age45-54.female', 'age45-54.male',
# 'age55-64.female', 'age55-64.male', 'age65-.female', 'age65-.male', ## youtube_channel_demographics

# 'averageViewPercentage', ## youtube_channel_locations

# 'annotationClickThroughRate',
# 'annotationCloseRate', 'averageViewDuration_y', 'comments', 'dislikes',
# 'estimatedMinutesWatched_y', 'likes', 'shares', 'subscribersGained',
# 'subscribersLost', 'views_y', 'redViews', 'estimatedRevenue',
# 'estimatedAdRevenue', 'estimatedRedPartnerRevenue', 'grossRevenue',
# 'cpm_y', 'monetizedPlaybacks', 'adImpressions', 'playbackBasedCpm', ## youtube_daily_channel_basics


# 'subscribers_gained', 'subscribers_lost', 'subscribers_count' ## youtube_subscriber

# 최종 컬럼 정리
users_col_need = [
    # 계정, 날짜 기준
    'youtube_user_id', 'date', 

    # 계정 정보
    'channel_id', 'channel_title_x', 'phone_num', 'report_user_id', 'published_at_x', 'viewCount', 'subscriberCount', 'videoCount', 'yt_search_keyword',

    # 참여도 관련
    'views_y','redViews','UNSUBSCRIBED', 'SUBSCRIBED', 'comments', 'likes', 'shares', 'dislikes', 'estimatedMinutesWatched_y', 'averageViewDuration_y', 'averageViewPercentage',

    # 수익 관련
    'estimated_revenue', 'estimatedRevenue', 'estimated_ad_revenue', 'estimatedAdRevenue', 'estimated_red_partner_revenue', 'estimatedRedPartnerRevenue', 
    'gross_revenue', 'grossRevenue', 'cpm_x', 'cpm_y', 'playback_based_cpm', 'playbackBasedCpm',

    # 구독자 관련
    'subscribers_count',
    'subscribersGained', 'subscribers_gained',
    'subscribersLost', 'subscribers_lost',

    # 광고 관련
    'ad_impressions', 'adImpressions',
    'monetized_playbacks', 'monetizedPlaybacks',

    # 연령 및 성별 관련
    'age13-17.female', 'age13-17.male',
    'age18-24.female', 'age18-24.male', 'age25-34.female', 'age25-34.male',
    'age35-44.female', 'age35-44.male', 'age45-54.female', 'age45-54.male',
    'age55-64.female', 'age55-64.male', 'age65-.female', 'age65-.male'
]
len(users_col_need)

# 최종 계정 분석 데이터셋

# 컬럼명 변경
merge_df_users_fin = merge_df_users[users_col_need]
merge_df_users_fin = merge_df_users_fin.rename(columns={'channel_title_x':'channel_title', 'published_at_x':'published_at', 'views_y':'views', 'cpm_x' : 'cpm',
                                                        'estimatedMinutesWatched_y':'estimatedMinutesWatched', 'averageViewDuration_y':'averageViewDuration'})

# 전처리
# 환율 적용 - 해당기간 평균환율 : 1322.42
exchange_rate_col = ['estimatedRevenue', 'estimatedAdRevenue', 'estimatedRedPartnerRevenue', 'grossRevenue', 'cpm_y', 'playbackBasedCpm']
merge_df_users_fin[exchange_rate_col] = merge_df_users_fin[exchange_rate_col] * 1322.42

# 시청 시간 단위 통일
merge_df_users_fin['averageViewDuration'] = round(merge_df_users_fin['averageViewDuration'] / 60, 2) ## 초 단위 분으로 변경

# null값 평균환율 적용값으로 대체
merge_df_users_fin['estimated_revenue'] = merge_df_users_fin['estimated_revenue'].fillna(merge_df_users_fin['estimatedRevenue'])
merge_df_users_fin['estimated_ad_revenue'] = merge_df_users_fin['estimated_ad_revenue'].fillna(merge_df_users_fin['estimatedAdRevenue'])
merge_df_users_fin['estimated_red_partner_revenue'] = merge_df_users_fin['estimated_red_partner_revenue'].fillna(merge_df_users_fin['estimatedRedPartnerRevenue'])
merge_df_users_fin['gross_revenue'] = merge_df_users_fin['gross_revenue'].fillna(merge_df_users_fin['grossRevenue'])
merge_df_users_fin['cpm'] = merge_df_users_fin['cpm'].fillna(merge_df_users_fin['cpm_y'])
merge_df_users_fin['playback_based_cpm'] = merge_df_users_fin['playback_based_cpm'].fillna(merge_df_users_fin['playbackBasedCpm'])

# null값 정확한 테이블값으로 대체
merge_df_users_fin['ad_impressions'] = merge_df_users_fin['ad_impressions'].fillna(merge_df_users_fin['adImpressions'])
merge_df_users_fin['subscribers_gained'] = merge_df_users_fin['subscribers_gained'].fillna(merge_df_users_fin['subscribersGained'])
merge_df_users_fin['subscribers_lost'] = merge_df_users_fin['subscribers_lost'].fillna(merge_df_users_fin['subscribersLost'])
merge_df_users_fin['monetized_playbacks'] = merge_df_users_fin['monetized_playbacks'].fillna(merge_df_users_fin['monetizedPlaybacks'])

# 대체된 값 제거
merge_df_users_fin = merge_df_users_fin.drop(['estimatedRevenue', 'estimatedAdRevenue', 'estimatedRedPartnerRevenue', 'grossRevenue', 'cpm_y', 'playbackBasedCpm',
                                              'adImpressions','subscribersGained','subscribersLost','monetizedPlaybacks'],axis=1)

# 나머지 null값 측정안된값으로 0으로 대체
merge_df_users_fin[merge_df_users_fin.columns[11:]] = merge_df_users_fin[merge_df_users_fin.columns[11:]].fillna(0)

# merge_df_users_fin.to_csv('C:/py_src/awake/data/merge_df_users_fin.csv', encoding='utf-8-sig',index=False)

#### 콘텐츠 분석 데이터셋

# # 불필요컬럼 제거 - 모두 0인 컬럼
# youtube_videos = youtube_videos.drop(['annotationClickThroughRate', 'annotationCloseRate',
#                                       'annotationImpressions', 'annotationClickableImpressions',
#                                       'annotationClosableImpressions', 'annotationClicks', 'annotationCloses'],axis=1)

youtube_videos = pd.read_csv(file_path + 'youtube_videos.csv')

# 최종 컬럼 정리
video_col_need = [
    # 콘텐츠, 날짜 기준
    'youtube_user_id', 'video', 'end_date',

    # 참여도 관련
    'views', 'redViews', 'comments', 'likes', 'dislikes', 'shares', 'estimatedMinutesWatched', 'estimatedRedMinutesWatched', 'averageViewDuration', 'averageViewPercentage', 
    'videosAddedToPlaylists','videosRemovedFromPlaylists',

    # 수익 관련
    'estimatedRevenue', 'estimatedAdRevenue', 'grossRevenue','estimatedRedPartnerRevenue', 'playbackBasedCpm', 'cpm',

    # 구독자 관련
    'subscribersGained', 'subscribersLost',

    # 광고 관련
    'monetizedPlaybacks', 'adImpressions',

    # 카드 관련
    'cardClickRate', 'cardTeaserClickRate','cardImpressions', 'cardTeaserImpressions', 'cardClicks', 'cardTeaserClicks'
]
len(video_col_need)

# 최종 콘텐츠 분석 데이터셋

# 전처리
# 환율 적용 - 해당기간 평균환율 : 1322.42
exchange_rate_col = ['estimatedRevenue', 'estimatedAdRevenue', 'estimatedRedPartnerRevenue', 'grossRevenue', 'cpm', 'playbackBasedCpm']
youtube_videos[exchange_rate_col] = youtube_videos[exchange_rate_col] * 1322.42

# 최종 데이터셋 정리
youtube_videos = youtube_videos[video_col_need].sort_values(['video', 'end_date']).reset_index(drop=True)

youtube_videos.head()

# youtube_videos.to_csv('C:/py_src/awake/data/youtube_videos.csv', encoding='utf-8-sig',index=False)
