==> 5.Modeling(4).ipynb <==
import pymongo
import pandas as pd
import numpy as np
import math
from datetime import datetime

# import dask.dataframe as dd ## 대규모 데이터 처리
# import vaex ## 대규모 데이터 처리

# from sklearn.linear_model import LinearRegression ## likes, dislikes 대체
from sklearn.preprocessing import StandardScaler
from sklearn.ensemble import IsolationForest
from sklearn.model_selection import train_test_split

from sklearn.ensemble import RandomForestClassifier
from sklearn.ensemble import RandomForestRegressor, GradientBoostingRegressor
from lightgbm import LGBMRegressor
import xgboost as xgb
from sklearn.model_selection import KFold, cross_val_score, train_test_split
from sklearn.metrics import mean_squared_error

from imblearn.under_sampling import RandomUnderSampler
from sklearn.metrics import accuracy_score, confusion_matrix, classification_report, roc_auc_score, roc_curve

import os
from dotenv import load_dotenv

file_path = 'C:/py_src/awake/data/'

# 계정 / 콘텐츠 분석 데이터셋 불러오기
merge_df_users_fin = pd.read_csv(file_path + 'merge_df_users_fin.csv', low_memory=False)
# youtube_videos = pd.read_csv(file_path + 'youtube_videos.csv')

## 계정 데이터 분석

### 데이터 분할
- 미래를 예측하기 위한 모델이므로 시간순으로 데이터 분할
- 시작날짜 : 2023-03-26
- 종료날짜 : 2024-05-06
- 전체 기간의 80% 날짜 : 2024-02-11

# 컬럼 정리
unique_col = merge_df_users_fin.columns[:11]
x_col = merge_df_users_fin.columns[11:].drop('estimated_ad_revenue') ## y값 제거

# 데이터 분할
train_data = merge_df_users_fin[merge_df_users_fin['date'] <= '2024-02-11']
test_data = merge_df_users_fin[merge_df_users_fin['date'] > '2024-02-11']

# 불균형 확인
print(train_data.shape)
print(test_data.shape)

### 모델 기법 적용

# 각 모델 정의
rf_model = RandomForestRegressor(random_state=42)
gbm_model = GradientBoostingRegressor(random_state=42)
lgbm_model = LGBMRegressor(random_state=42)
xgb_model = xgb.XGBRegressor(random_state=42)

# 모델 리스트
models = {
    'RandomForest': rf_model,
    'GBM': gbm_model,
    'LightGBM': lgbm_model,
    'XGBoost': xgb_model
}

# # 모델별 교차 검증 결과 저장
# results = {}

# for model_name, model in models.items():
#     print(f"\n{model_name} 모델 성능 평가 중...")
    
#     # 교차 검증
#     cv_scores = cross_val_score(model, train_data[x_col], train_data['subscribers_count'], cv=5, scoring='neg_mean_squared_error')
    
#     # 평균 RMSE 계산
#     rmse_scores = np.sqrt(-cv_scores)  # neg_mean_squared_error는 음수이므로 양수로 변환 후 제곱근
#     mean_rmse = rmse_scores.mean()
#     print(f"{model_name} 교차 검증 평균 RMSE: {mean_rmse}")
    
#     # 결과 저장
#     results[model_name] = mean_rmse

### 모델 성능 평가

from sklearn.ensemble import RandomForestRegressor
from sklearn.metrics import mean_squared_error, r2_score
import numpy as np

for model in models:
    # 모델 정의 및 학습
    models[model].fit(train_data[x_col], train_data['estimated_ad_revenue'])

    # 예측
    y_pred = models[model].predict(test_data[x_col])

    # MSE 계산
    mse = mean_squared_error(test_data['estimated_ad_revenue'], y_pred)

    # RMSE 계산
    rmse = np.sqrt(mse)

    # R² 값 계산
    r2 = r2_score(test_data['estimated_ad_revenue'], y_pred)

    # Adjusted R² 계산
    n = len(test_data)  # 샘플 수
    p = test_data.shape[1]  # 독립 변수(특성) 수
    adjusted_r2 = 1 - ((1 - r2) * (n - 1) / (n - p - 1))

    # 결과 출력
    print(model)
    print(f"R² 값: {r2:.4f}")
    print(f"Adjusted R² 값: {adjusted_r2:.4f}")
    print(f"MSE: {mse:.4f}")
    print(f"RMSE: {rmse:.4f}")
    print("----------------------------------------")
    print("")

# # 비선형모델 활용 변수 선정
# importances_df = pd.DataFrame({
#     'features': x_col,
#     'rf_importance': models['RandomForest'].feature_importances_,
#     'gbm_importance': models['GBM'].feature_importances_,
#     'lgbm_importance': models['LightGBM'].feature_importances_,
#     'xgb_importance': models['XGBoost'].feature_importances_    
# })
# importances_df['lgbm_importance'] = importances_df['lgbm_importance'] / importances_df['lgbm_importance'].sum() ## 다른 모델과 중요도 단위 다르기 때문에 정규화
# importances_df['mean_importance'] = importances_df[['rf_importance', 'gbm_importance', 'lgbm_importance', 'xgb_importance']].mean(axis=1) ## 각 변수별 모델 중요도 평균

### 모델 성능 개선

스태킹

from sklearn.ensemble import StackingRegressor
from sklearn.linear_model import Ridge

# 스태킹 모델 정의
stacking_model = StackingRegressor(
    estimators=[
    ('random_forest', models['RandomForest']),
    ('gbm', models['GBM']),
    ('lightgbm', models['LightGBM']),
    ('xgboost', models['XGBoost'])
    ],
    final_estimator=Ridge()
)

# 스태킹 모델 학습
stacking_model.fit(train_data[x_col], train_data['estimated_ad_revenue'])

# 예측
y_pred = stacking_model.predict(test_data[x_col])

# 성능 평가
# R² 값 계산
r2 = r2_score(test_data['estimated_ad_revenue'], y_pred)

# Adjusted R² 계산
n = len(test_data)  # 샘플 수
p = test_data.shape[1]  # 독립 변수(특성) 수
adjusted_r2 = 1 - ((1 - r2) * (n - 1) / (n - p - 1))

# MSE 계산
mse = mean_squared_error(test_data['estimated_ad_revenue'], y_pred)

# RMSE 계산
rmse = mean_squared_error(test_data['estimated_ad_revenue'], y_pred, squared=False)

print(f"스태킹 앙상블 모델 R2: {r2:.4f}")
print(f"스태킹 앙상블 모델 Adjusted R2: {adjusted_r2:.4f}")
print(f"스태킹 앙상블 모델 MSE: {mse:.4f}")
print(f"스태킹 앙상블 모델 RMSE: {rmse:.4f}")

### 실제데이터 결과 확인

# 예측 결과 확인
y_pred = stacking_model.predict(merge_df_users_fin[x_col])
merge_df_users_fin['predict'] = y_pred

# 계정별 구독자수 평균, 구독자 예측수 평균 비교
result_df = merge_df_users_fin[['youtube_user_id','date','channel_title','estimated_ad_revenue','predict']]

# 1개월, 3개월, 6개월, 1년 뒤 구독자수 예측값
result_df['date'] = pd.to_datetime(result_df['date']) 

result_df['1_month_future_predict'] = result_df.groupby('youtube_user_id')['predict'].shift(-30).rolling(window=30).sum()
result_df['3_month_future_predict'] = result_df.groupby('youtube_user_id')['predict'].shift(-90).rolling(window=90).sum()
result_df['6_month_future_predict'] = result_df.groupby('youtube_user_id')['predict'].shift(-180).rolling(window=180).sum()
result_df['12_month_future_predict'] = result_df.groupby('youtube_user_id')['predict'].shift(-365).rolling(window=365).sum()

# 계정별 기간에 따른 광고 수익 예측값 비교
# result_df_final = result_df.groupby(['youtube_user_id'])[['estimated_ad_revenue','predict','1_month_future_predict','3_month_future_predict','6_month_future_predict','12_month_future_predict']].sum().reset_index()

# 계정별로 최종 평균값을 계산
result_df_final = result_df.groupby('youtube_user_id').agg({
    'estimated_ad_revenue': 'mean', ## 1일 평균 광고수익
    'predict': 'mean', ## 1일 평균 광고수익 예측
    '1_month_future_predict': 'mean',
    '3_month_future_predict': 'mean',
    '6_month_future_predict': 'mean',
    '12_month_future_predict': 'mean'
}).reset_index()

result_df[result_df['youtube_user_id'] == '63d2239450eb530dfd137d1e']

result_df_final[result_df_final['predict'] > 0].sort_values('predict')

result_df[result_df['youtube_user_id'].isin(result_df_final[result_df_final['predict'] < 0]['youtube_user_id'])]['channel_title'].unique()

# 상위, 하위 20개 계정
top_user_id_20_6_month = result_df_final[result_df_final['6_month_future_predict'] > 0].sort_values(['6_month_future_predict'],ascending=False)['youtube_user_id'].iloc[:20]
bottom_user_id_20_6_month = result_df_final[result_df_final['6_month_future_predict'] > 0].sort_values(['6_month_future_predict'],ascending=False)['youtube_user_id'].iloc[-20:]

print(len(result_df[result_df['youtube_user_id'].isin(top_user_id_20_6_month)]['channel_title'].unique()))
print(result_df[result_df['youtube_user_id'].isin(top_user_id_20_6_month)]['channel_title'].unique())

print(len(result_df[result_df['youtube_user_id'].isin(bottom_user_id_20_6_month)]['channel_title'].unique()))
print(result_df[result_df['youtube_user_id'].isin(bottom_user_id_20_6_month)]['channel_title'].unique())

len(set(top_user_id_20_6_month) & set(top_user_id_20_6_month) & set(top_user_id_20_6_month))

## 콘텐츠 데이터 분석

# 최종 콘텐츠 분석 데이터셋
# 환율 적용 - 해당기간 평균환율 : 1322.42
exchange_rate_col = ['estimatedRevenue', 'estimatedAdRevenue', 'estimatedRedPartnerRevenue', 'grossRevenue', 'cpm', 'playbackBasedCpm']
youtube_videos[exchange_rate_col] = youtube_videos[exchange_rate_col] * 1322.42

# 잘못된값 처리
youtube_videos['estimatedRevenue'] = np.where(youtube_videos['estimatedRevenue'] < 0,
                                               youtube_videos['estimatedAdRevenue'] + youtube_videos['estimatedRedPartnerRevenue'],
                                               youtube_videos['estimatedRevenue'])

# 버그로 사용된 수치값 대체
youtube_videos['likes'] = np.where(youtube_videos['likes'] < 0, 0, youtube_videos['likes'])
youtube_videos['dislikes'] = np.where(youtube_videos['dislikes'] < 0, 0, youtube_videos['dislikes'])

### 중요지표확인

youtube_videos.columns

# y값 설정
# youtube_videos['estimatedRevenue']
# youtube_videos['net_subscribers_change']
# youtube_videos['engage_rate']
youtube_videos['net_subscribers_change'] = youtube_videos['subscribersGained'] - youtube_videos['subscribersLost']
youtube_videos['engage_rate'] = (youtube_videos['likes'] + youtube_videos['comments'] + youtube_videos['shares']) / youtube_videos['views']

# 조회수당 수익
youtube_videos['revenue_per_view'] = youtube_videos['estimatedRevenue'] / youtube_videos['views']

# 구독자당 수익
youtube_videos['revenue_per_Subscriber'] = youtube_videos['estimatedRevenue'] / youtube_videos['subscribersGained']

# YouTube Premium 수익
youtube_videos['estimatedRedPartnerRevenue']

# 수익 다변화 비율
youtube_videos['revenue_diversification_ratio'] = (youtube_videos['grossRevenue'] - youtube_videos['estimatedRevenue']) / youtube_videos['estimatedRevenue']

# 구독자 증가율
youtube_videos['subscriber_growth_rate'] = youtube_videos['subscribersGained'] / (youtube_videos['subscribersGained'] + youtube_videos['subscribersLost'])

# 구독자 감소율
youtube_videos['subscriber_loss_rate'] = youtube_videos['subscribersLost'] / (youtube_videos['subscribersGained'] + youtube_videos['subscribersLost'])

# 구독자 유지율
youtube_videos['subscriber_retention_rate'] = (youtube_videos['subscribersGained'] - youtube_videos['subscribersLost']) / youtube_videos['subscribersGained']

# 콘텐츠당 구독자 증가율
youtube_videos['subscriber_gain_per_content'] = youtube_videos['subscribersGained']/ youtube_videos['videosAddedToPlaylists']

# 구독자당 시청 시간
youtube_videos['watch_time_per_subscriber'] = youtube_videos['estimatedMinutesWatched'] / youtube_videos['subscribersGained']

# 광고재생률
youtube_videos['ad_playback_rate'] = youtube_videos['monetizedPlaybacks'] / youtube_videos['adImpressions']

상관분석

corr_df = youtube_videos[youtube_videos.columns[3:]].corr()
corr_df[['estimatedRevenue','net_subscribers_change','engage_rate']]

### 기간별 피처 생성

# 주별 피처 생성(콘텐츠별로 계산)
youtube_videos['weekly_revenue_per_view_std'] = youtube_videos.groupby('video')['revenue_per_view'].transform(lambda x: x.rolling(window=7).std())
youtube_videos['weekly_revenue_per_Subscriber_std'] = youtube_videos.groupby('video')['revenue_per_Subscriber'].transform(lambda x: x.rolling(window=7).std())
youtube_videos['weekly_estimatedRedPartnerRevenue_std'] = youtube_videos.groupby('video')['estimatedRedPartnerRevenue'].transform(lambda x: x.rolling(window=7).std())
youtube_videos['weekly_revenue_diversification_ratio_std'] = youtube_videos.groupby('video')['revenue_diversification_ratio'].transform(lambda x: x.rolling(window=7).std())
youtube_videos['weekly_subscriber_growth_rate_std'] = youtube_videos.groupby('video')['subscriber_growth_rate'].transform(lambda x: x.rolling(window=7).std())
youtube_videos['weekly_subscriber_loss_rate_std'] = youtube_videos.groupby('video')['subscriber_loss_rate'].transform(lambda x: x.rolling(window=7).std())
youtube_videos['weekly_subscriber_retention_rate_std'] = youtube_videos.groupby('video')['subscriber_retention_rate'].transform(lambda x: x.rolling(window=7).std())
youtube_videos['weekly_subscriber_gain_per_content_std'] = youtube_videos.groupby('video')['subscriber_gain_per_content'].transform(lambda x: x.rolling(window=7).std())
youtube_videos['weekly_watch_time_per_subscriber_std'] = youtube_videos.groupby('video')['watch_time_per_subscriber'].transform(lambda x: x.rolling(window=7).std())
youtube_videos['weekly_ad_playback_rate_std'] = youtube_videos.groupby('video')['ad_playback_rate'].transform(lambda x: x.rolling(window=7).std())

youtube_videos['weekly_revenue_per_view_trd'] = youtube_videos.groupby('video')['revenue_per_view'].transform(lambda x: (x - x.shift(7)) / x.shift(7))
youtube_videos['weekly_revenue_per_Subscriber_trd'] = youtube_videos.groupby('video')['revenue_per_Subscriber'].transform(lambda x: (x - x.shift(7)) / x.shift(7))
youtube_videos['weekly_estimatedRedPartnerRevenue_trd'] = youtube_videos.groupby('video')['estimatedRedPartnerRevenue'].transform(lambda x: (x - x.shift(7)) / x.shift(7))
youtube_videos['weekly_revenue_diversification_ratio_trd'] = youtube_videos.groupby('video')['revenue_diversification_ratio'].transform(lambda x: (x - x.shift(7)) / x.shift(7))
youtube_videos['weekly_subscriber_growth_rate_trd'] = youtube_videos.groupby('video')['subscriber_growth_rate'].transform(lambda x: (x - x.shift(7)) / x.shift(7))
youtube_videos['weekly_subscriber_loss_rate_trd'] = youtube_videos.groupby('video')['subscriber_loss_rate'].transform(lambda x: (x - x.shift(7)) / x.shift(7))
youtube_videos['weekly_subscriber_retention_rate_trd'] = youtube_videos.groupby('video')['subscriber_retention_rate'].transform(lambda x: (x - x.shift(7)) / x.shift(7))
youtube_videos['weekly_subscriber_gain_per_content_trd'] = youtube_videos.groupby('video')['subscriber_gain_per_content'].transform(lambda x: (x - x.shift(7)) / x.shift(7))
youtube_videos['weekly_watch_time_per_subscriber_trd'] = youtube_videos.groupby('video')['watch_time_per_subscriber'].transform(lambda x: (x - x.shift(7)) / x.shift(7))
youtube_videos['weekly_ad_playback_rate_trd'] = youtube_videos.groupby('video')['ad_playback_rate'].transform(lambda x: (x - x.shift(7)) / x.shift(7))

# 월별 피처 생성(콘텐츠별로 계산)
youtube_videos['monthly_revenue_per_view_std'] = youtube_videos.groupby('video')['revenue_per_view'].transform(lambda x: x.rolling(window=30).std())
youtube_videos['monthly_revenue_per_Subscriber_std'] = youtube_videos.groupby('video')['revenue_per_Subscriber'].transform(lambda x: x.rolling(window=30).std())
youtube_videos['monthly_estimatedRedPartnerRevenue_std'] = youtube_videos.groupby('video')['estimatedRedPartnerRevenue'].transform(lambda x: x.rolling(window=30).std())
youtube_videos['monthly_revenue_diversification_ratio_std'] = youtube_videos.groupby('video')['revenue_diversification_ratio'].transform(lambda x: x.rolling(window=30).std())
youtube_videos['monthly_subscriber_growth_rate_std'] = youtube_videos.groupby('video')['subscriber_growth_rate'].transform(lambda x: x.rolling(window=30).std())
youtube_videos['monthly_subscriber_loss_rate_std'] = youtube_videos.groupby('video')['subscriber_loss_rate'].transform(lambda x: x.rolling(window=30).std())
youtube_videos['monthly_subscriber_retention_rate_std'] = youtube_videos.groupby('video')['subscriber_retention_rate'].transform(lambda x: x.rolling(window=30).std())
youtube_videos['monthly_subscriber_gain_per_content_std'] = youtube_videos.groupby('video')['subscriber_gain_per_content'].transform(lambda x: x.rolling(window=30).std())
youtube_videos['monthly_watch_time_per_subscriber_std'] = youtube_videos.groupby('video')['watch_time_per_subscriber'].transform(lambda x: x.rolling(window=30).std())
youtube_videos['monthly_ad_playback_rate_std'] = youtube_videos.groupby('video')['ad_playback_rate'].transform(lambda x: x.rolling(window=30).std())

youtube_videos['monthly_revenue_per_view_trd'] = youtube_videos.groupby('video')['revenue_per_view'].transform(lambda x: (x - x.shift(30)) / x.shift(30))
youtube_videos['monthly_revenue_per_Subscriber_trd'] = youtube_videos.groupby('video')['revenue_per_Subscriber'].transform(lambda x: (x - x.shift(30)) / x.shift(30))
youtube_videos['monthly_estimatedRedPartnerRevenue_trd'] = youtube_videos.groupby('video')['estimatedRedPartnerRevenue'].transform(lambda x: (x - x.shift(30)) / x.shift(30))
youtube_videos['monthly_revenue_diversification_ratio_trd'] = youtube_videos.groupby('video')['revenue_diversification_ratio'].transform(lambda x: (x - x.shift(30)) / x.shift(30))
youtube_videos['monthly_subscriber_growth_rate_trd'] = youtube_videos.groupby('video')['subscriber_growth_rate'].transform(lambda x: (x - x.shift(30)) / x.shift(30))
youtube_videos['monthly_subscriber_loss_rate_trd'] = youtube_videos.groupby('video')['subscriber_loss_rate'].transform(lambda x: (x - x.shift(30)) / x.shift(30))
youtube_videos['monthly_subscriber_retention_rate_trd'] = youtube_videos.groupby('video')['subscriber_retention_rate'].transform(lambda x: (x - x.shift(30)) / x.shift(30))
youtube_videos['monthly_subscriber_gain_per_content_trd'] = youtube_videos.groupby('video')['subscriber_gain_per_content'].transform(lambda x: (x - x.shift(30)) / x.shift(30))
youtube_videos['monthly_watch_time_per_subscriber_trd'] = youtube_videos.groupby('video')['watch_time_per_subscriber'].transform(lambda x: (x - x.shift(30)) / x.shift(30))
youtube_videos['monthly_ad_playback_rate_trd'] = youtube_videos.groupby('video')['ad_playback_rate'].transform(lambda x: (x - x.shift(30)) / x.shift(30))

# null값 대체
youtube_videos = youtube_videos.fillna(0) ## NaN
youtube_videos = youtube_videos.replace([np.inf, -np.inf], 0) ## inf

# youtube_videos.to_csv('C:/py_src/awake/data/youtube_videos_credit.csv', encoding='utf-8-sig', index=False)

# youtube_videos = dd.read_csv('C:/py_src/awake/data/youtube_videos_credit.csv')
youtube_videos = pd.read_csv(file_path + 'youtube_videos_credit.csv')

### 데이터 분할
- 미래를 예측하기 위한 모델이므로 시간순으로 데이터 분할
- 시작날짜 : 2023-03-26
- 종료날짜 : 2024-05-06
- 전체 기간의 80% 날짜 : 2024-02-11

# 컬럼 정리
unique_col = youtube_videos.columns[:3]
remove_col = ['estimatedRevenue','estimatedAdRevenue','grossRevenue','net_subscribers_change','subscribersGained','subscribersLost','engage_rate','views']
x_col = youtube_videos.columns.drop(list(unique_col) + remove_col) ## y값 제거, y값 변수 관여 변수 제거

# 데이터 분할
train_data = youtube_videos[youtube_videos['end_date'] <= '2024-02-11']
test_data = youtube_videos[youtube_videos['end_date'] > '2024-02-11']

# 불균형 확인
print(train_data.shape)
print(test_data.shape)

### 모델 기법 적용

# 각 모델 정의
# rf_model = RandomForestRegressor(random_state=42)
# gbm_model = GradientBoostingRegressor(random_state=42)
# lgbm_model = LGBMRegressor(random_state=42)
xgb_model = xgb.XGBRegressor(random_state=42)

# 모델 리스트
models = {
    # 'RandomForest': rf_model,
    # 'GBM': gbm_model,
    # 'LightGBM': lgbm_model,
    'XGBoost': xgb_model
}

# # 모델별 교차 검증 결과 저장
# results = {}

# for model_name, model in models.items():
#     print(f"\n{model_name} 모델 성능 평가 중...")
    
#     # 교차 검증
#     cv_scores = cross_val_score(model, train_data[x_col], train_data['subscribers_count'], cv=5, scoring='neg_mean_squared_error')
    
#     # 평균 RMSE 계산
#     rmse_scores = np.sqrt(-cv_scores)  # neg_mean_squared_error는 음수이므로 양수로 변환 후 제곱근
#     mean_rmse = rmse_scores.mean()
#     print(f"{model_name} 교차 검증 평균 RMSE: {mean_rmse}")
    
#     # 결과 저장
#     results[model_name] = mean_rmse

### 모델 성능 평가

feature_importances_dict = {}

from sklearn.ensemble import RandomForestRegressor
from sklearn.metrics import mean_squared_error, r2_score
import numpy as np

for target in ['estimatedRevenue','net_subscribers_change','engage_rate']:
    # 모델 정의 및 학습
    models['XGBoost'].fit(train_data[x_col], train_data[target]) ## 'estimatedRevenue','net_subscribers_change','engage_rate

    # 변수 중요도
    feature_importances_dict[target] = models['XGBoost'].feature_importances_

    # 예측
    y_pred = models['XGBoost'].predict(test_data[x_col])

    # MSE 계산
    mse = mean_squared_error(test_data[target], y_pred)

    # RMSE 계산
    rmse = np.sqrt(mse)

    # R² 값 계산
    r2 = r2_score(test_data[target], y_pred)

    # Adjusted R² 계산
    n = len(test_data)  # 샘플 수
    p = test_data.shape[1]  # 독립 변수(특성) 수
    adjusted_r2 = 1 - ((1 - r2) * (n - 1) / (n - p - 1))

    # 결과 출력
    print(target)
    print(f"R² 값: {r2:.4f}")
    print(f"Adjusted R² 값: {adjusted_r2:.4f}")
    print(f"MSE: {mse:.4f}")
    print(f"RMSE: {rmse:.4f}")
    print("----------------------------------------")
    print("")

### 변수 중요도 확인
- 각 y값에 대한 중요도 평균

importances_df = pd.DataFrame({
    'features': x_col,
    'estimatedRevenue_importance': feature_importances_dict['estimatedRevenue'],
    'net_subscribers_change_importance': feature_importances_dict['net_subscribers_change'],
    'engage_rate_importance': feature_importances_dict['engage_rate']
    })

importances_df['mean_importance'] = importances_df[['estimatedRevenue_importance', 'net_subscribers_change_importance', 'engage_rate_importance']].mean(axis=1)

최종 지표 선정

# 신용평가 최종 지표 선정
final_col = list(importances_df.sort_values(['mean_importance'],ascending=False).iloc[:20]['features'])
final_col

# # 신용평가 최종 지표
# final_col = ['dislikes', 'estimatedRedPartnerRevenue', 'likes', 'redViews', 'monthly_revenue_diversification_ratio_trd', 'shares', 'revenue_diversification_ratio', 'estimatedMinutesWatched',
#              'revenue_per_view', 'comments', 'averageViewDuration', 'playbackBasedCpm', 'averageViewPercentage', 'monthly_watch_time_per_subscriber_std', 'watch_time_per_subscriber',
#              'adImpressions', 'monetizedPlaybacks', 'monthly_estimatedRedPartnerRevenue_std', 'subscriber_gain_per_content', 'videosRemovedFromPlaylists']

# # 신용 평가 최종 지표 활용 모델 성능 확인
# from sklearn.ensemble import RandomForestRegressor
# from sklearn.metrics import mean_squared_error, r2_score
# import numpy as np

# for target in ['estimatedRevenue','net_subscribers_change','engage_rate']:
#     # 모델 정의 및 학습
#     models['XGBoost'].fit(train_data[final_col], train_data[target]) ## 'estimatedRevenue','net_subscribers_change','engage_rate

#     # 변수 중요도
#     feature_importances_dict[target] = models['XGBoost'].feature_importances_

#     # 예측
#     y_pred = models['XGBoost'].predict(test_data[final_col])

#     # MSE 계산
#     mse = mean_squared_error(test_data[target], y_pred)

#     # RMSE 계산
#     rmse = np.sqrt(mse)

#     # R² 값 계산
#     r2 = r2_score(test_data[target], y_pred)

#     # Adjusted R² 계산
#     n = len(test_data)  # 샘플 수
#     p = test_data.shape[1]  # 독립 변수(특성) 수
#     adjusted_r2 = 1 - ((1 - r2) * (n - 1) / (n - p - 1))

#     # 결과 출력
#     print(target)
#     print(f"R² 값: {r2:.4f}")
#     print(f"Adjusted R² 값: {adjusted_r2:.4f}")
#     print(f"MSE: {mse:.4f}")
#     print(f"RMSE: {rmse:.4f}")
#     print("----------------------------------------")
#     print("")

### 가중치 설정
- 최종 지표 활용

importances_df_final = importances_df[importances_df['features'].isin(final_col)][['features','mean_importance']].reset_index(drop=True).sort_values('mean_importance', ascending=False)
importances_df_final['weight'] = importances_df_final['mean_importance'] / importances_df_final['mean_importance'].sum() ## 가중치 설정
importances_df_final

평가요소 분류

# 상환이력
eval_col1 = ['dislikes','likes','shares','comments','redViews']
importances_df_final[importances_df_final['features'].isin(eval_col1)]['weight'].sum()

# 부채수준
eval_col2 = ['estimatedRedPartnerRevenue','monthly_estimatedRedPartnerRevenue_std','revenue_per_view']
importances_df_final[importances_df_final['features'].isin(eval_col2)]['weight'].sum()

# 신용거래기간
eval_col3 = ['subscriber_gain_per_content','videosRemovedFromPlaylists','monthly_watch_time_per_subscriber_std','watch_time_per_subscriber']
importances_df_final[importances_df_final['features'].isin(eval_col3)]['weight'].sum()

# 신용형태
eval_col4 = ['revenue_diversification_ratio','monthly_revenue_diversification_ratio_trd','playbackBasedCpm','monetizedPlaybacks','adImpressions']
importances_df_final[importances_df_final['features'].isin(eval_col4)]['weight'].sum()

# 비금융/마이데이터
eval_col5 = ['averageViewDuration','averageViewPercentage','estimatedMinutesWatched']
importances_df_final[importances_df_final['features'].isin(eval_col5)]['weight'].sum()

가중치 적용

# 가중치 실제값 적용
# 상환이력
youtube_videos['score1'] = (youtube_videos[eval_col1] * np.array(importances_df_final[importances_df_final['features'].isin(eval_col1)]['weight'])).sum(axis=1)

# 부채수준
youtube_videos['score2'] = (youtube_videos[eval_col2] * np.array(importances_df_final[importances_df_final['features'].isin(eval_col2)]['weight'])).sum(axis=1)

# 신용거래기간
youtube_videos['score3'] = (youtube_videos[eval_col3] * np.array(importances_df_final[importances_df_final['features'].isin(eval_col3)]['weight'])).sum(axis=1)

# 신용형태
youtube_videos['score4'] = (youtube_videos[eval_col4] * np.array(importances_df_final[importances_df_final['features'].isin(eval_col4)]['weight'])).sum(axis=1)

# 비금융/마이데이터
youtube_videos['score5'] = (youtube_videos[eval_col5] * np.array(importances_df_final[importances_df_final['features'].isin(eval_col5)]['weight'])).sum(axis=1)

# 계정별 신용점수 확인 - 계정별 평가요소 합계
credit_df = youtube_videos.groupby('youtube_user_id')[['score1','score2','score3','score4','score5']].sum().reset_index()

# 불필요 및 이상치 제거
credit_df = credit_df[~credit_df['youtube_user_id'].isin(['639bb8dcd603b8138e33780b'])].reset_index(drop=True)
# '639bb8dcd603b8138e33780b' ## 없는 계정 및 이상치

from sklearn.preprocessing import MinMaxScaler

# 1. MinMaxScaler 적용
scaler = MinMaxScaler()
# credit_df['score_scale'] = scaler.fit_transform(credit_df[['score']])
credit_df['score1_scale'] = scaler.fit_transform(credit_df[['score1']])
credit_df['score2_scale'] = scaler.fit_transform(credit_df[['score2']])
credit_df['score3_scale'] = scaler.fit_transform(credit_df[['score3']])
credit_df['score4_scale'] = scaler.fit_transform(credit_df[['score4']])
credit_df['score5_scale'] = scaler.fit_transform(credit_df[['score5']])

# credit_df['score_final'] = scaler.fit_transform(credit_df[['score_scale']]) * 1000
# credit_df['score1_final'] = scaler.fit_transform(credit_df[['score1_scale']]) * 350 # 평가요소 0.284 * 0 이 아닌 비율 0.12 --> 0.03 : 0.075 * 0.35 --> 0.03  : 0.15
# credit_df['score2_final'] = scaler.fit_transform(credit_df[['score2_scale']]) * 280 # 평가요소 0.245 * 0 이 아닌 비율 0.46 --> 0.11 : 0.27 * 0.28 --> 0.076 : 0.38
# credit_df['score3_final'] = scaler.fit_transform(credit_df[['score3_scale']]) * 120 # 평가요소 0.123 * 0 이 아닌 비율 0.25 --> 0.03 : 0.075 * 0.12 --> 0.009 : 0.04
# credit_df['score4_final'] = scaler.fit_transform(credit_df[['score4_scale']]) * 180 # 평가요소 0.275 * 0 이 아닌 비율 0.58 --> 0.16 : 0.4 * 0.18 --> 0.072 : 0.36
# credit_df['score5_final'] = scaler.fit_transform(credit_df[['score5_scale']]) * 80 # 평가요소 0.073 * 0 이 아닌 비율 0.9 --> 0.07 : 0.18 * 0.08 --> 0.014 : 0.07

credit_df['score1_final'] = scaler.fit_transform(credit_df[['score1_scale']]) * 430
credit_df['score2_final'] = scaler.fit_transform(credit_df[['score2_scale']]) * 410
credit_df['score3_final'] = scaler.fit_transform(credit_df[['score3_scale']]) * 50
credit_df['score4_final'] = scaler.fit_transform(credit_df[['score4_scale']]) * 60
credit_df['score5_final'] = scaler.fit_transform(credit_df[['score5_scale']]) * 50

# 평가요소 분류 활용 스코어링
credit_df['credit_score'] = credit_df[['score1_final','score2_final','score3_final','score4_final','score5_final']].sum(axis=1)

# 계정 확인 테이블
user_info_df = merge_df_users_fin[['youtube_user_id','channel_title']].drop_duplicates().reset_index(drop=True)
user_info_df = user_info_df[user_info_df['channel_title']!='0'].drop_duplicates().reset_index(drop=True)
user_info_df = user_info_df[~user_info_df['channel_title'].isnull()].reset_index(drop=True)
user_info_df = user_info_df.drop_duplicates(['youtube_user_id']).reset_index(drop=True)
user_info_df = user_info_df.drop_duplicates(['channel_title']).reset_index(drop=True)

# 신용평가점수 테이블
credit_df_fin = pd.merge(credit_df[['youtube_user_id','credit_score']],user_info_df,how='left',on='youtube_user_id')
credit_df_fin = credit_df_fin[['youtube_user_id','channel_title','credit_score']]
credit_df_fin = credit_df_fin[~credit_df_fin['channel_title'].isnull()].reset_index(drop=True)

credit_df_fin.sort_values(['credit_score'],ascending=False)

### 변동계수
- y값 활용 : 'estimatedRevenue', 'net_subscribers_change', 'engage_rate'
- 0값이 많아 왜곡된 변동계수를 계산할 수 있음 --> 0값 제거후 변동계수 산출

# 각 y값의 0비율 확인
# 'estimatedRevenue', 'net_subscribers_change', 'engage_rate'
print('estimatedRevenue의 0비율 : ', round((youtube_videos['estimatedRevenue']==0).sum() / len(youtube_videos),2))
print('net_subscribers_change의 0비율 : ', round((youtube_videos['net_subscribers_change']==0).sum() / len(youtube_videos),2))
print('engage_rate의 0비율 : ', round((youtube_videos['engage_rate']==0).sum() / len(youtube_videos),2))

# estimatedRevenue
youtube_videos['weekly_estimatedRevenue_cv'] = youtube_videos.groupby('video')['estimatedRevenue'].transform(lambda x: x.rolling(window=7).std() / x.rolling(window=7).mean())
youtube_videos['monthly_estimatedRevenue_cv'] = youtube_videos.groupby('video')['estimatedRevenue'].transform(lambda x: x.rolling(window=30).std() / x.rolling(window=30).mean())

# net_subscribers_change
youtube_videos['weekly_net_subscribers_change_cv'] = youtube_videos[youtube_videos['net_subscribers_change']!=0].groupby('video')['net_subscribers_change'].transform(lambda x: x.rolling(window=7).std() / x.rolling(window=7).mean())
youtube_videos['monthly_net_subscribers_change_cv'] = youtube_videos[youtube_videos['net_subscribers_change']!=0].groupby('video')['net_subscribers_change'].transform(lambda x: x.rolling(window=30).std() / x.rolling(window=30).mean())

# engage_rate
youtube_videos['weekly_engage_rate_cv'] = youtube_videos[youtube_videos['engage_rate']!=0].groupby('video')['engage_rate'].transform(lambda x: x.rolling(window=7).std() / x.rolling(window=7).mean())
youtube_videos['monthly_engage_rate_cv'] = youtube_videos[youtube_videos['engage_rate']!=0].groupby('video')['engage_rate'].transform(lambda x: x.rolling(window=30).std() / x.rolling(window=30).mean())

cv_col = ['weekly_estimatedRevenue_cv','monthly_estimatedRevenue_cv','weekly_net_subscribers_change_cv','monthly_net_subscribers_change_cv','weekly_engage_rate_cv','monthly_engage_rate_cv']

# null값 대체
youtube_videos = youtube_videos.fillna(0) ## NaN
youtube_videos = youtube_videos.replace([np.inf, -np.inf], 0) ## inf

# 계정별 변동계수 평균
coefvar_df = youtube_videos.groupby('youtube_user_id')[cv_col].mean().reset_index()

# 최종 신용평가 테이블
credit_coef_df = pd.merge(credit_df_fin, coefvar_df, how='left', on='youtube_user_id')
# credit_df_final = credit_df_final[~credit_df_final['channel_title'].isnull()].reset_index(drop=True)

# 가중치 역으로 활용하여 최종 변동 계수 도출
# (1 - 0.36) + (1 - 0.3) + (1 - 0.34) 활용 스케일링
# 0.32 / 0.35 / 0.33 가중치 적용
credit_coef_df['cv_score'] = (
                              (((credit_coef_df['weekly_estimatedRevenue_cv'] + credit_coef_df['monthly_estimatedRevenue_cv']) / 2) * 0.32) + 
                              (((credit_coef_df['weekly_net_subscribers_change_cv'] + credit_coef_df['monthly_net_subscribers_change_cv']) / 2) * 0.35) + 
                              (((credit_coef_df['weekly_engage_rate_cv'] + credit_coef_df['monthly_engage_rate_cv']) / 2) * 0.33)
                              )

# 최종 데이터셋 확인
credit_df_final = credit_coef_df[['youtube_user_id','channel_title','credit_score','cv_score']]

credit_df_final.sort_values('credit_score',ascending=False)

### 실제데이터 결과 확인

credit_df_final[credit_df_final['credit_score'] >= 400]

부도 가능성 매우 안전

credit_df_final[credit_df_final['youtube_user_id']=='63d77c9650eb530dfd139f8b']

부도 가능성 안전

credit_df_final[credit_df_final['youtube_user_id']=='63eb4f87ee122e631992279f']

부도 가능성 보통

credit_df_final[credit_df_final['youtube_user_id']=='6401e117d746c60e1271fdef']

부도 가능성 위험

credit_df_final[credit_df_final['youtube_user_id'].isin(['62d11f080b4c4c7502a5be3d', '64399d0b659261656b3f0681', 
                                                         '6444cfbea224e30e2de2b264', '6486b00519c22b644dded32c', 
                                                         '649b0fb29247f326464c53f6', '64da8a9ef638790e0f74bae7', 
                                                         '64dc7156f638790e0f74d1a2'])]

부도 가능성 고위험

credit_df_final.sort_values(['cv_score'])

credit_df_final[(credit_df_final['credit_score'] < 200) | (credit_df_final['cv_score'] >= 0.3)]
