==> 8.Distribution.ipynb <==
패키지 로드

import pymongo
import pandas as pd
import numpy as np
from sklearn.preprocessing import StandardScaler
from sklearn.ensemble import IsolationForest
import xgboost as xgb
import os
from dotenv import load_dotenv
import joblib
import cloudpickle

import warnings
warnings.filterwarnings('ignore')

유튜브 분석 피클 로드

# 불러오기
with open('youtube_analysis.pkl', 'rb') as f:
    youtube_analysis = cloudpickle.load(f)

분석에 필요한 데이터 로드

# 필요데이터 경로
need_data_path = 'C:/py_src/awake/data/need/'

# 필요데이터 로드
# 변수명, 파일명 변경하면 실행안됨(주의!)
user_info_df, trained_data, features_weight, input_data = youtube_analysis.data_preparation_need(need_data_path)

MongDB 연동

# MongoDB 연동
# 각자 ID, PW 입력
db = youtube_analysis.mongodb_connection('external_user','nrAx9lyGH5K7lbgv')

데이터 준비

youtube_user_df = youtube_analysis.data_preparation_user(db)
youtube_contents_df = youtube_analysis.data_preparation_contents(db)

분석 결과 확인

# Topic 1 - 콘텐츠 영향력 데이터 내 Fraud 색출을 통한 채널 진성 영향력 파악
result_topic1_df = youtube_analysis.result_topic1()

# Topic 2 - 구독자 수 예측
result_topic2_df = youtube_analysis.result_topic2()

# Topic 3 - 광고수익 예측
result_topic3_df = youtube_analysis.result_topic3()

# Topic 4 - 신용평가 스코어링, 부도예측확률 분포도 구조 개발
result_topic4_df = youtube_analysis.result_topic4()
